refactoring folder structure to support FE
