+++ examples/json/grpc-to-rest/flogo.json
@@ -9,7 +9,7 @@
     {
       "name": "flogo-grpc",
       "id": "MyProxy",
-      "ref": "github.com/project-flogo/grpc/trigger",
+      "ref": "github.com/project-flogo/grpc/trigger/grpc",
       "settings": {
         "port": "9096",
         "protoName": "petstore"
