+++ support/grpcservicegen.go
@@ -116,7 +116,7 @@ import (
 	{{end}}
 	"log"
 	"errors"
-	servInfo "github.com/project-flogo/grpc/trigger"
+	servInfo "github.com/project-flogo/grpc/trigger/grpc"
 	"github.com/golang/protobuf/jsonpb"
 	"google.golang.org/grpc"
 )
