+++ trigger/grpc/trigger.go
@@ -1,4 +1,4 @@
-package trigger
+package grpc
 
 import (
 	"context"
